+++ HPL/sources/resources/ResourceBase.cpp
@@ -31,7 +31,7 @@ namespace hpl {
 	//////////////////////////////////////////////////////////////////////////
 
 	//-----------------------------------------------------------------------
-	iResourceBase::iResourceBase(tString asName,unsigned long alPrio){
+	iResourceBase::iResourceBase(tString asName,unsigned int alPrio){
 		mlTime = (unsigned long)time(NULL);
 		mlPrio = alPrio;
 		mlHandle = 0;
